Role checking is now done with middleware and  Extract admin-specific activites to Admin\JobController from JobController

+++ routes/api.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\JobController;
 use App\Http\Controllers\UserController;
+use App\Http\Controllers\WorkProfileController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -15,14 +16,25 @@ Route::prefix('register')->group(function () {
     Route::post('job-seeker', [AuthController::class, 'jobSeekerRegister']);
 });
 
-Route::post('login', [AuthController::class, 'login']);
+Route::post('login', [AuthController::class, 'login'])->name('login');
 
-Route::post('logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
+Route::post('logout', [AuthController::class, 'logout'])->middleware('auth:sanctum')->name('logout');
 
-Route::apiResource('jobs', JobController::class);
+Route::apiResource('jobs', JobController::class)->only(['index', 'show']);
 
-Route::post('/jobs/{id}/restore', [JobController::class, 'restore']);
-Route::post('/jobs/{id}/force-delete', [JobController::class, 'forceDelete']);
+Route::middleware(['auth:sanctum', 'recruiter'])->group(function () {
+    Route::prefix('recruiter')->name('recruiter.')->group(function () {
+        Route::apiResource('jobs', JobController::class)->except(['index', 'show']);
+        Route::get('/jobs/mine', [JobController::class, 'myJobs'])->name('jobs.mine');
+    });
+});
 
-Route::get('/my-jobs', [JobController::class, 'myJobs'])->middleware('auth:sanctum');
-Route::get('/my-deleted-jobs', [JobController::class, 'myDeletedJobs'])->middleware('auth:sanctum');
+Route::prefix('admin')->middleware(['auth:sanctum', 'admin'])->group(function () {
+    Route::get('jobs', [\App\Http\Controllers\Admin\JobController::class, 'index'])->name('jobs.index');
+    Route::post('/jobs/{id}/restore', [\App\Http\Controllers\Admin\JobController::class, 'restore'])->name('jobs.restore');
+    Route::post('/jobs/{id}/force-delete', [\App\Http\Controllers\Admin\JobController::class, 'forceDelete'])->name('jobs.forceDelete');
+});
+
+
+
+Route::get('/admin/work-profile', [WorkProfileController::class, 'index'])->middleware('auth:sanctum');
